Version 2.1
Improved the startup time and memory usage.  Automatically turn the status line on when displaying the tag name in the status line.  Both of those features were contributed by Keith Reynolds.  Changed the types of tags displayed for C and vim files.  Added updating of the ctags database when a file is written.  Mostly fixed a problem where the script caused the cursor to not remember its column position as it was moved vertically.  Details on the script page.